feat: Add Edit & Delete dialogs + search feature

+++ jobtrail-backend/src/routes/applications/route.ts
@@ -1,7 +1,6 @@
 import Elysia from "elysia";
 import {
     postApplicationSchema, getApplicationSchema, putApplicationSchema, deleteApplicationsSchema,
-    getApplicationsSchema
 } from "./schema";
 import { db } from "../../db/db";
 import { applicationsTable } from "../../db/schema";
@@ -9,12 +8,27 @@ import { eq } from "drizzle-orm";
 import { getClaims } from "../../utils/auth/getClaims"
 import { getApplication } from "../../utils/applications"
 import { Application } from "./types"
+import { StatusCodes } from "http-status-codes";
+
+const validate = async (
+    id: number,
+    authorization: string,
+    set: { status?: number | string }
+) => {
+    const application = await getApplication(id)
+    const claims = await getClaims(authorization)
+
+    if(claims.sub != application.userId) {
+        set.status = StatusCodes.UNAUTHORIZED
+        return "Unauthorized"
+    }
+}
 
 export const applicationRouter = new Elysia({ prefix: "/applications" })
     // @ts-ignore
     .onBeforeHandle(async({ set, headers: { authorization }, route}) => {
         if(!authorization) {
-            set.status = 401
+            set.status = StatusCodes.UNAUTHORIZED
             return "Unauthorized"
         }
     })
@@ -24,7 +38,7 @@ export const applicationRouter = new Elysia({ prefix: "/applications" })
         const { companyName, email, applicationStatus, position, content } = body
 
         if(!authorization) {
-            set.status = 401
+            set.status = StatusCodes.UNAUTHORIZED
             return
         }
 
@@ -43,10 +57,9 @@ export const applicationRouter = new Elysia({ prefix: "/applications" })
                 content: content
             })
 
-            set.status = 201
+            set.status = StatusCodes.CREATED
 
     }, postApplicationSchema)
-
     // @ts-ignore
     .get("/",async({jwt, headers: { authorization }}) => {
         // const claims: ClaimTypes = await jwt.verify(auth.value)
@@ -79,28 +92,23 @@ export const applicationRouter = new Elysia({ prefix: "/applications" })
                 acceptedCount: applications.filter(app => app.applicationStatus == "ACCEPTED").length,
             }
         }
-    }, getApplicationsSchema)
-
-
-    .onBeforeHandle(async({set, params, headers: { authorization }}) => {
-        const applicationId = Number(params.id)
-        const application = await getApplication(applicationId)
-        const claims = await getClaims(authorization!)
-
-        if(claims.sub != application.userId) {
-            set.status = 401
-            return "Unauthorized"
-        }
     })
-    .get("/:id", async({ params, set }) => {
+
+
+    .get("/:id", async({ params, set, headers: { authorization } }) => {
         const id = Number(params.id)
+
+        const unauthorized = await validate(id, authorization!, set)
+        if(unauthorized) {
+            return unauthorized
+        }
 
         const result = await db.select()
             .from(applicationsTable)
             .where(eq(applicationsTable.id, id))
        
         if(result.length == 0) {
-            set.status = 404
+            set.status = StatusCodes.NOT_FOUND
             throw `Application with id: ${id} not found`
         }
 
@@ -117,22 +125,32 @@ export const applicationRouter = new Elysia({ prefix: "/applications" })
     }, getApplicationSchema)
 
 
-    .put("/:id", async({ params, body, set }) => {
+    .put("/:id", async({ params, body, set, headers: { authorization } }) => {
         const id = Number(params.id)
+
+        const unauthorized = await validate(id, authorization!, set)
+        if(unauthorized) {
+            return unauthorized
+        }
 
         await db.update(applicationsTable)
             .set(body)
             .where(eq(applicationsTable.id, id))
 
-        set.status = 204
+        set.status = StatusCodes.NO_CONTENT
 
     }, putApplicationSchema)
 
 
-    .delete("/:id", async({params, set}) => {
+    .delete("/:id", async({params, set, headers: { authorization }}) => {
         const id = Number(params.id)
+
+        const unauthorized = await validate(id, authorization!, set)
+        if(unauthorized) {
+            return unauthorized
+        }
         
         await db.delete(applicationsTable).where(eq(applicationsTable.id, id))
 
-        set.status = 204
+        set.status = StatusCodes.NO_CONTENT
     }, deleteApplicationsSchema)
